Change Screen & New FeedReader class
Change Screen now works

The FeedReader class reader the JSonf feed every X seconds. Then create
the balloon object corresponding to whats need to be displayed, then
compare the newly created balloon with the wan't currently displayed.
Balloon present in the feed but not on the server are add via a
NewBalloon Message. Balloon present in the server but not in the feed
are deleted from the server (and screens) with a PopBalloonMessage