Always draw balloons on top.
Fixed an issue whereby a balloon would be drawn at the same time as its label, causing the last drawn balloon to always be the last thing on top. Changed so that content boxes are drawn first and then the balloons are drawn. Unfortunately turns out the box was being drawn half way up the string so I changed the position and cropped the images to be the right size.